+++ src/aqua_diagnostics/__init__.py
@@ -1,11 +1,12 @@
 from .teleconnections import Teleconnection
 from .timeseries import GregoryPlot, SeasonalCycle, Timeseries
-from .ensemble import EnsembleTimeseries, EnsembleLatLon, EnsembleZonal
 from .global_biases import GlobalBiases
 from .radiation import Radiation
-
-__all__ = ["Teleconnection", "GregoryPlot", "SeasonalCycle", "Timeseries", "GlobalBiases", "Radiation", "EnsembleTimeseries", "EnsembleLatLon", "EnsembleZonal"]
-
-
-
-
+from .ensemble import EnsembleTimeseries, EnsembleLatLon, EnsembleZonal
+from .ecmean import PerformanceIndices, GlobalMean
+__all__ = ["Teleconnection",
+           "GregoryPlot", "SeasonalCycle", "Timeseries", 
+           "GlobalBiases",
+           "Radiation", 
+           "EnsembleTimeseries", "EnsembleLatLon", "EnsembleZonal",
+           "GlobalMean", "PerformanceIndices"]